Move ConsultationDeskReportsNumberCard to components/index/cards
